Updating UI automation test
