+++ demo/src/androidTest/java/com/google/android/fhir/demo/uitest/pages/AddPatientPage.kt
@@ -1,0 +1,47 @@
+/*
+ * Copyright 2021 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.android.fhir.demo.uitest.pages
+
+import androidx.test.espresso.Espresso.closeSoftKeyboard
+import androidx.test.espresso.Espresso.onView
+import androidx.test.espresso.action.ViewActions.click
+import androidx.test.espresso.assertion.ViewAssertions.matches
+import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
+import androidx.test.espresso.matcher.ViewMatchers.withId
+import androidx.test.espresso.matcher.ViewMatchers.withText
+import com.google.android.fhir.demo.R
+import org.hamcrest.Matchers.allOf
+
+class AddPatientPage {
+  /*Add patient page objects*/
+  private val pageName = "Add Patient"
+  private val addPatientButton = R.id.add_patient
+
+  fun validate_page() {
+    /*Validate Add patient page name*/
+    onView(
+      allOf(
+      withText(R.string.add_patient), withText(pageName))).check(matches(isDisplayed()))
+    closeSoftKeyboard()
+    Thread.sleep(3000)
+  }
+
+  fun clickOnAddPatientButton() {
+    onView(withId(addPatientButton)).perform(click())
+  }
+
+}
