+++ demo/src/androidTest/java/com/google/android/fhir/demo/uitest/tests/BaseTest.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2021 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.android.fhir.demo.uitest.tests
+
+import androidx.test.ext.junit.rules.ActivityScenarioRule
+import com.google.android.fhir.demo.MainActivity
+import org.junit.Rule
+
+
+open class BaseTest {
+  @get:Rule
+  val activityRule: ActivityScenarioRule<MainActivity> =
+    ActivityScenarioRule(MainActivity::class.java)
+}
