+++ demo/src/main/res/layout/patient_list_view.xml
@@ -24,23 +24,21 @@
 
     <TextView
         android:id="@+id/patient_count"
-        android:text="0 Patients"
-        android:layout_centerHorizontal="true"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-    />
+        android:layout_centerHorizontal="true"
+        android:text="0 Patients" />
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/patient_list"
-        android:layout_below="@+id/patient_count"
         android:name="com.google.fhirengine.example.PatientListFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
+        android:layout_below="@+id/patient_count"
         android:layout_marginLeft="16dp"
         android:layout_marginRight="16dp"
         android:scrollbars="vertical"
         app:layoutManager="LinearLayoutManager"
         tools:context=".PatientListActivity"
-        tools:listitem="@layout/patient_list_item_view"
-    />
+        tools:listitem="@layout/patient_list_item_view" />
 </RelativeLayout>
